move build tests into separate file
